+++ .github/scripts/scorecard/ajv-bundle.js
@@ -1,3 +1,4 @@
+// @ts-nocheck
 "use strict";
 var __getOwnPropNames = Object.getOwnPropertyNames;
 var __commonJS = (cb, mod) => function __require() {